fixed: fixed a few bugs tied to deck editing and saving
Fixed a bug when you attempt to open a deck that does not belong to you.
Previously, the app would redirect to a user that included another
users id and the related deck id. When we copy a deck, a new deck id
should be generated, since we are creating a brand new deck and this
should put the user on the deckbuilder page.

Fixed a bug tied to deleting decks were the decks were not being removed
from Firebase do to an issue with the updateSave logic. Previously, the
logic would check the local state decks for the deck id that was deleted
and if not found, just return early. Since a deck was being deleted,
the deck id would never be found. Now, the app will lookup that deck id
from the database, and then attempt to delete that deck.

Fixed a bug tied to saving changes to a deck were the confirmation modal
would automatically open on the profile deck view page. Previously, if you
saved a change to the deck in the deckbuilder view, if you then switched
to your profile view and clicked on that same deck, the confirmation
modal would automatically open and warn you about saving changes to your
deck. This has been fixed by making sure the component is disposed of once
you leave the deckbuilder view.

Finally, fixed a bug when attempting to delete a deck from the database.
Previously, the firestore rules would not allow this action.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>