[GH-44] added: missing card images for death moth and space portal
Added missing card images for the Death Moth (card 47) from the
Lobby Levels set, and Space Portal (card 66) from the Lobby
Levels set.

Added new promo card for September 2025, the Phase Plate.

Replaced a few card images for a few promos were the card sizes
did not align with the rest of the images in the app.

Fixes: https://github.com/scottwestover/backrooms-tcg-deck-builder/issues/24
Fixes: https://github.com/scottwestover/backrooms-tcg-deck-builder/issues/25

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>