[GH-47] fixed bug with export deck to file
Fixed bug with the export deck to file feature that
was broken as part of the changes introduced for
the deck randomization feature.

The issue was tied to trying to lazy load the import
which broke the feature. For the time being, the
change was reverted and the lazy loading can be
reviewed at a later time.

fixes: https://github.com/scottwestover/backrooms-tcg-deck-builder/issues/47

Signed-off-by: Scott Westover <swestover@manh.com>